pyproject.toml
[tool.poetry]
name = "autogen-jira"
version = "0.1.0"
description = ""
authors = ["lanbaoshen <lanbaoshen@icloud.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
atlassian-python-api = "^3.41.19"
pydantic-settings = "^2.7.1"
loguru = "^0.7.3"
autogen-agentchat = "^0.4.3"
autogen-ext = {extras = ["azure", "openai"], version = "^0.4.3"}
fastapi = {extras = ["standard"], version = "^0.115.7"}
sqlmodel = "^0.0.22"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
